Enhanced support for unmocking inherited subroutines (RT77439)
After unmocking a subroutine that has been inherited from
another module, calling that subroutine in the child module
will dispatch to the parent module's subroutine, no matter
what changes have happend to the parent module in the mean time.

Additionally, original() will dispatch to the parent module's
subroutine, if the subroutine was inherited at the time of mocking.
If it was not, original() will continue to return the original
subroutine as normal.